+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Changed
+
+- Implement `Clone` for `teloxide::RequestError`
+  - `RequestError::Network` now accepts `Arc<reqwest::Error>` [**BC**]
+  - `RequestError::InvalidJson` now accepts `source: Arc<serde_json::Error>` [**BC**]
+  - `RequestError::Io` now accepts `Arc<io::Error>` [**BC**]
+
 ## 0.14.1 - 2025-03-30
 
 ### Fixed